--- src/features/projectsV2/ProjectListControls/utils.tsx
@@ -1,11 +1,0 @@
-import { TreeProjectClassification } from '@planet-sdk/common';
-
-export const availableFilters: TreeProjectClassification[] = [
-  'large-scale-planting',
-  'agroforestry',
-  'natural-regeneration',
-  'managed-regeneration',
-  'urban-planting',
-  'mangroves',
-  'other-planting',
-];
