+++ src/features/projectsV2/ProjectsMap/MultipleProjectsView.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useMemo } from 'react';
-import { useProjects } from '../../ProjectsContext';
-import ProjectMarkers, { CategorizedProjects } from '../ProjectMarkers';
-import { getProjectCategory } from '../utils';
-import { zoomOutMap } from '../../../../utils/mapsV2/zoomToProjectSite';
-import { SetState } from '../../../common/types/common';
+import { useProjects } from '../ProjectsContext';
+import ProjectMarkers, { CategorizedProjects } from './ProjectMarkers';
+import { getProjectCategory } from '../../../utils/projectV2';
+import { zoomOutMap } from '../../../utils/mapsV2/zoomToProjectSite';
+import { SetState } from '../../common/types/common';
 import { ViewState } from 'react-map-gl-v7';
-import { MapRef } from '../../../../utils/mapsV2/zoomToProjectSite';
+import { MapRef } from '../../../utils/mapsV2/zoomToProjectSite';
 interface MultipleProjectsViewProps {
   setViewState: SetState<ViewState>;
   mapRef: MapRef;
