+++ src/features/projectsV2/ProjectsMap/ProjectMarkers/ProjectMarkerIcon.tsx
@@ -9,7 +9,7 @@ import {
   UrbanRestoration,
 } from '../../../../../public/assets/images/icons/myForestMapIcons/PointMarkerIcons';
 import { MapProjectProperties } from '../../../common/types/projectv2';
-import { getProjectCategory } from '../utils';
+import { getProjectCategory } from '../../../../utils/projectV2';
 import themeProperties from '../../../../theme/themeProperties';
 
 type Props = {
