--- src/features/projectsV2/ProjectsMap/utils.tsx
@@ -1,20 +1,0 @@
-import {
-  ExtendedProject,
-  MapProjectProperties,
-} from '../../common/types/projectv2';
-
-export const getProjectCategory = (
-  projectProperties: MapProjectProperties | ExtendedProject
-) => {
-  if (
-    projectProperties.purpose === 'trees' &&
-    projectProperties.isTopProject &&
-    projectProperties.isApproved
-  ) {
-    return 'topProject';
-  } else if (projectProperties.allowDonations) {
-    return 'regularProject';
-  } else {
-    return 'nonDonatableProject';
-  }
-};
